Fixed timestamp

+++ mqtt2influxdb/mqtt2influxdb.py
@@ -58,7 +58,7 @@ def on_message(client, userdata, msg):
         json_body = [
           {
             "measurement": "tempout",
-            "time": time.time() * 1000000000,
+            "time": long(time.time() * 1000000000),
             "fields": {
               "tempout": pl["temperature"]
             }
